feat!: refactor conditions into a module
