+++ src/conditions/const.ts
@@ -1,0 +1,10 @@
+import { ChainId } from '../types';
+
+export const SUPPORTED_CHAINS = [
+  ChainId.MAINNET,
+  ChainId.GOERLI,
+  ChainId.POLYGON,
+  ChainId.MUMBAI,
+];
+
+export const USER_ADDRESS_PARAM = ':userAddress';
