+++ src/conditions/predefined/index.ts
@@ -1,0 +1,1 @@
+export { ERC721Balance, ERC721Ownership } from './erc721';
